Step 1: Starting WebShop Extension Scenario
This commit contains a project frame which is a good starting point for
any SAPUI5 based HANA Cloud Application development. It is used here as
starting point for the WebShop Extension Scenario.
It already consists of the default structure of a typical SAPUI5 Maven
project and by very small UI5 examples it is demonstrated how to read
and write data from backend systems by means of OData Services

The corresponding WebShop Extension Scenario description on SCN
http://www.sdn.sap.com/irj/scn/index?rid=/library/uuid/8016f858-369d-3010-1f9c-9402c4ecad3fis
a step by step tutorial for any developers. 
In Chapter 2 it is described how the project frame is build up and how
it runs.

Used technologies SAPUI5, OData, SAP HANA Cloud Connectivity